requirements.txt
Flask
openai
Pillow
python-dotenv
SpeechRecognition
PyPDF2
Werkzeug
Requests
google-generativeai
nest-asyncio